requirements.txt
MarkupSafe==1.0
absl-py==0.5.0
asn1crypto==1.3.0
astor==0.7.1
backports.weakref==1.0rc1
bleach==1.5.0
blinker==1.4
boto==2.49.0
boto3==1.9.17
botocore==1.12.18
bz2file==0.98
cachetools==4.0.0
certifi==2018.8.24
cffi==1.14.0
chardet==3.0.4
ChatterBot==0.8.7
chatterbot-corpus==1.1.4
Click==7.0
cryptography==2.8
docutils==0.14
Flask==0.12.2
Flask-WTF==0.14.2
future==0.17.1
gast==0.2.0
gensim==3.6.0
google-api-core==1.16.0
google-auth==1.12.0
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.16.1
h5py==2.8.0
html5lib==0.9999999
idna==2.7
itsdangerous==0.24
jieba==0.39
Jinja2==2.10
jmespath==0.9.3
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
mathparse==0.1.2
mkl-service==2.3.0
mock==4.0.1
mongoengine==0.15.3
nltk==3.3
numpy==1.15.2
oauthlib==2.1.0
pandas==0.23.4
protobuf==3.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pymongo==3.7.1
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dateutil==2.6.1
python-twitter==3.5
pytz==2018.5
PyYAML==3.13
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
rsa==4.0
s3transfer==0.1.13
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
smart-open==1.10.0
SQLAlchemy==1.2.13
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
urllib3==1.23
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.12.1
WTForms==2.2.1
xlrd==1.1.0
